+++ CHANGELOG.md
@@ -39,8 +39,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- **CUMULUS-2439**
+  - Added CMR search client setting to the CreateReconciliationReport lambda function.
+  - Added `cmr_search_client_config` tfvars to the archive and cumulus terraform modules.
+  - Updated CreateReconciliationReport lambda to search CMR collections with CMRSearchConceptQueue.
+- **CUMULUS-2638**
+  - Adds documentation to clarify bucket config object use.
+
 ### Changed
 
+- **CUMULUS-2638**
+  - Transparent to users, remove typescript type `BucketType`.
 - **CUMULUS-2718**
   - Updated config for SyncGranules to support optional `workflowStartTime`
   - Updated SyncGranules to provide `createdAt` on output based on `workflowStartTime` if provided,